Add user relation with task and restrict acces to task by user owner

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { FiltersDto } from './dto/filters-dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,40 +15,40 @@ export class TasksService {
   ){}
 
 
-  async getTasks(filters: FiltersDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filters);
+  async getTasks(filters: FiltersDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filters, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const foundTask = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const foundTask = await this.taskRepository.findOne({where: {id, userId: user.id}});
     if (!foundTask) {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
     return foundTask;
   }
 
-  async deleteTask(id: number): Promise<void> {
+  async deleteTask(id: number, user: User): Promise<void> {
     /* One way, this way make 2 query to the database, first veify that exist the task and before deleted */
     /* const foundTask: Task = await this.getTaskById(id);
     foundTask.remove(); */
 
     /* Second way, faster, make only one request to database */
-    const result = await this.taskRepository.delete(id);
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
     
   }
 
-  async updateTaskStatus(id: number, staus:TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, staus:TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = staus;
     await task.save();
     return task;
   }
   
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
 }
